new method

+++ src/main/java/sem2/sep2/server/view/ManageRoomViewController.java
@@ -102,33 +102,53 @@ public class ManageRoomViewController implements ViewController{
     //button here rg1
     public void AddingButtonPressed(ActionEvent actionEvent) throws SQLException
     {
-        String id = room_id.getText();
-        String Price = price.getText();
-        String type = roomType.getTypeSelector();
-        if(viewModel.isIntDouble(id)&&viewModel.isIntDouble(Price)){
-            int roomId = Integer.parseInt(id);
-            Double price_ = Double.parseDouble(Price);
-            roomDao.createRoom(new Room(roomId,type,price_));
-            viewModel.alarm("Successfully","Add successfully!");
-        }else{
-            viewModel.alarm("Alarm","Please check the id and price.");
+        if(viewModel.confirm()){
+            String id = room_id.getText();
+            String Price = price.getText();
+            String type = (String) roomType.getValue();
+            String available = (String) availability.getValue();
+            if(viewModel.isIntDouble(id)&&viewModel.isIntDouble(Price)){
+                int roomId = Integer.parseInt(id);
+                if(!roomDao.isRoomExist(roomId)){
+                    Double price_ = Double.parseDouble(Price);
+                    roomDao.createRoom(new Room(roomId,type,price_,available));
+                    viewModel.alarm("Successfully","Add successfully!");
+                }else{
+                    viewModel.alarm("Alarm","The room has already exist.");
+                }
+            }else{
+                viewModel.alarm("Alarm","Please check the id and price.");
+            }
+            refresh();
+            reset();
         }
-        refresh();
-        reset();
+
     }
 
     public void EditingButtonPressed(ActionEvent actionEvent)throws SQLException
     {
-        String id = room_id.getText();
-        String Price = price.getText();
-        String type = roomType.getTypeSelector();
-        if(viewModel.isIntDouble(id)&&viewModel.isIntDouble(Price)){
-
-        }else{
-            viewModel.alarm("Alarm","Please check the id and price.");
+        if(viewModel.confirm()){
+            String id = room_id.getText();
+            String Price = price.getText();
+            String type = (String) roomType.getValue();
+            String available = (String) availability.getValue();
+            if(viewModel.isIntDouble(id)&&viewModel.isIntDouble(Price)){
+                int roomId = Integer.parseInt(id);
+                if(roomDao.isRoomExist(roomId)){
+                    Double price_ = Double.parseDouble(Price);
+                    roomDao.deleteRoom(roomId);
+                    roomDao.createRoom(new Room(roomId,type,price_,available));
+                    viewModel.alarm("Successfully","Editing room successfully.");
+                }else{
+                    viewModel.alarm("Alarm","The room does not exist");
+                }
+            }else{
+                viewModel.alarm("Alarm","Please check the id and price.");
+            }
+            refresh();
+            reset();
         }
-        refresh();
-        reset();
+
     }
 
 
@@ -138,10 +158,14 @@ public class ManageRoomViewController implements ViewController{
         String id = idnumber.getText();
         if(viewModel.isIntDouble(id)){
             int roomId = Integer.parseInt(id);
-            if(roomDao.deleteRoom(roomId)){
-                viewModel.alarm("Successfully","Delete Successfully");
+            if(roomDao.isRoomExist(roomId)){
+                if(roomDao.deleteRoom(roomId)){
+                    viewModel.alarm("Successfully","Delete Successfully");
+                }else{
+                    viewModel.alarm("Failed","Delete Failed");
+                }
             }else{
-                viewModel.alarm("Failed","Delete Failed");
+                viewModel.alarm("Failed","The room doesn't exist");
             }
         }else{
             viewModel.alarm("Alarm","Please check the id.");
